Added 3 new components, created logic for login
I've added 3 new component: "LoginForm, GoalBar, ProgressCircle" to which I've also created some logic. For the login logic was all placed inside the "loginForm", first checking the input data using a form resolver and "useForm" . Then I've upgraded it to use a better form resolver using "yup", Then I've introduced a boolean which can be turned on and off for whether the user wants to register or login. Finally, my progressCircle component has a logic built in to calculate the percentage of how many hours of sleep have been achieved, and a circle which fills up depending on that amount.